cast to avoid MacOS compiler complaint
Comparing this ssize_t value vs sizeof(buf) is correct --
we've already handled any negative values -- but the Apple
compiler hates it, so cast.

+++ backend/fs.c
@@ -1524,7 +1524,7 @@ fs_readlink(void *softc __unused, struct l9p_request *req)
 	linklen = readlink(file->name, buf, sizeof(buf));
 	if (linklen < 0)
 		error = errno;
-	else if (linklen >= sizeof(buf))
+	else if ((size_t)linklen >= sizeof(buf))
 		error = ENOMEM; /* todo: allocate dynamically */
 	else if ((req->lr_resp.rreadlink.target = strndup(buf,
 	    (size_t)linklen)) == NULL)
